requirements.txt
watchdog
python-dotenv
numpy
requests
beautifulsoup4
python-magic-bin
pdfminer.six
python-docx
tqdm
google-generativeai
dotenv